+++ src/NetworkFramework/edge_field.jl
@@ -1,7 +1,12 @@
 """
-Expand to a web field from a matrix of raw values.
+Expand to a web field from a vector of raw values.
 """
 abstract type EdgeFieldRawBlueprint <: Blueprint end
+
+"""
+Expand to a web field from a matrix.
+"""
+abstract type EdgeFieldMatrixBlueprint <: Blueprint end
 
 """
 Expand to a web field from an adjacency list.
@@ -43,25 +48,24 @@ function define_reflexive_web_field_component(
         quote
             module $Value_
             import EcologicalNetworksDynamics: F, NF, D, Brought
-            const Web = $mod.$Web
+            const ew = $ew
+            const Web = D.component(ew)
             const _Web = typeof(Web)
             const d = $d
-            const (web, field) = D.content(d)
+            const T = $T
             end
         end
     ).args |> last)
 
     #---------------------------------------------------------------------------------------
     # From raw values.
+
     bpmod.eval(quote
         mutable struct Raw <: NF.EdgeFieldRawBlueprint
-            $field::Vector{$T}
-            $web::Brought(Web)
-            Raw($field, $web) = new(NF.construct(d, Raw, $field), $web)
-            Raw($field; $web = _Web) = Raw($field, $web)
+            $short::Vector{T}
+            Raw($short) = new(NF.construct(d, Raw, $short))
         end
-        NF.data(bp::Raw) = bp.$field
-        F.implied_blueprint_for(bp::Raw, ::_Web) = NF.implied_web(d, Web, bp)
+        NF.data(bp::Raw) = bp.$short
         F.early_check(bp::Raw) = NF.early_check(d, bp)
         F.late_check(model, bp::Raw, data) = NF.late_check(d, model, bp, data)
         F.expand!(model, bp::Raw, data) = NF.expand!(d, model, bp, data)
@@ -70,17 +74,38 @@ function define_reflexive_web_field_component(
     end)
 
     #---------------------------------------------------------------------------------------
+    # From a matrix.
+
+    sparse = D.is_sparse(d)
+    M = sparse ? SparseMatrix : Matrix
+    bpmod.eval(
+        quote
+            mutable struct Matrix <: NF.EdgeFieldMatrixBlueprint
+                $short::$M{T}
+                Matrix($short) = new(NF.construct(d, Matrix, $short))
+            end
+            NF.data(bp::Matrix) = bp.$short
+            F.implied_blueprint_for(bp::Matrix, ::_Web) = NF.implied_web(d, Web, bp)
+            F.early_check(bp::Matrix) = NF.early_check(d, bp)
+            F.late_check(model, bp::Matrix, data) = NF.late_check(d, model, bp, data)
+            F.expand!(model, bp::Matrix, data) = NF.expand!(d, model, bp, data)
+            NF.define_blueprint(Matrix, $"a $(sparse ? "sparse " : "")matrix")
+            export Matrix
+        end,
+    )
+
+    #---------------------------------------------------------------------------------------
     # From an adjacency list.
 
     bpmod.eval(
         quote
             mutable struct Adjacency <: NF.EdgeFieldAdjacencyBlueprint
-                $field::NF.Adjacency{$T}
+                $short::NF.Adjacency{T}
                 $web::Brought(Web) # Not exactly useful. Keep for consistency.
-                Adjacency($field, $web) = new(NF.construct(d, Adjacency, $field), $web)
-                Adjacency($field; $web = _Web) = Adjacency($field, $web)
+                Adjacency($short, $web) = new(NF.construct(d, Adjacency, $short), $web)
+                Adjacency($short; $web = _Web) = Adjacency($short, $web)
             end
-            NF.data(bp::Adjacency) = bp.$field
+            NF.data(bp::Adjacency) = bp.$short
             F.implied_blueprint_for(bp::Adjacency, ::_Web) = NF.implied_web(d, Web, bp)
             F.early_check(bp::Adjacency) = NF.early_check(d, bp)
             F.late_check(model, bp::Adjacency, data) = NF.late_check(d, model, bp, data)
@@ -97,9 +122,9 @@ function define_reflexive_web_field_component(
         bpmod.eval(
             quote
                 mutable struct Flat <: NF.EdgeFieldFlatBlueprint
-                    $field::$T
+                    $short::T
                 end
-                NF.data(bp::Flat) = bp.$field
+                NF.data(bp::Flat) = bp.$short
                 F.early_check(bp::Flat) = NF.early_check(d, bp)
                 F.late_check(model, bp::Flat, data) = NF.late_check(d, model, bp, data)
                 F.expand!(model, bp::Flat, data) = NF.expand!(d, model, bp, data)
@@ -127,15 +152,15 @@ function define_reflexive_web_field_component(
     mod.eval(
         quote
             $D.component(::$DT) = $comp
-            (::$_Value)($field, args...; kwargs...) =
-                $construct($d, $Value, $field, args...; kwargs...)
+            (::$_Value)($short, args...; kwargs...) =
+                $construct($d, $Value, $short, args...; kwargs...)
         end,
     )
 
     if may_flat(d)
         R = flat(d) # Receiver type.
         mod.eval(quote
-            (::$_Value)($field::$R) = $Value.Flat($field)
+            (::$_Value)($short::$R) = $Value.Flat($short)
         end)
     end
 
@@ -168,4 +193,113 @@ function define_reflexive_web_field_component(
 
     comp
 
+end
+
+# ==========================================================================================
+# Extracted logic + extension points.
+# Mostly duplicated from node field.
+
+#-------------------------------------------------------------------------------------------
+# Contextless check.
+
+check(d::EdgeField, value) = inputconvert(D.type(d), value)
+
+const EdgeIndex = Tuple{Int,Int}
+const EdgeLabel = Tuple{Symbol,Symbol}
+
+check_with_ref(d::EdgeField, value, (i, j)::EdgeIndex) =
+    try
+        check(d, value)
+    catch e
+        e isa F.InputError || rethrow(e)
+        with_context!(e, "On edge [$i, $j]")
+    end
+check_with_ref(d::EdgeField, value, (a, b)::EdgeLabel) =
+    try
+        check(d, value)
+    catch e
+        e isa F.InputError || rethrow(e)
+        with_context!(e, "On edge $(repr(a)) => $(repr(b))")
+    end
+# Useful when constructing from raw values because edge indices are unknown yet.
+check_with_raw_ref(d::EdgeField, value, i::Int) =
+    try
+        check(d, value)
+    catch e
+        e isa F.InputError || rethrow(e)
+        with_context!(e, "At raw edge value [$i]")
+    end
+
+#-------------------------------------------------------------------------------------------
+# Model-aware checks.
+
+function check_with_ref(d::EdgeField, against::Model, value, i::EdgeIndex, l::EdgeLabel)
+    try
+        value = check(d, value)
+        check(d, against, value, i, l)
+    catch e
+        e isa F.InputError || rethrow(e)
+        i, j = i
+        a, b = l
+        with_context!(e, "On edge $(repr(a)) => $(repr(b)) ([$i, $j])")
+    end
+end
+
+# Extract indexes for both incident classes.
+function indexes(d::EdgeField, network::Network)
+    web = D.web(d)
+    N.index.(network, D.sidenames(web))
+end
+
+# Infer any reference type from the other one.
+function check_with_ref(d::EdgeField, against, value, (i, j)::EdgeIndex)
+    model = get_model(against)
+    network = NF.network(model)
+    src, tgt = indexes(d, network)
+    (a, b) = N.to_label.((src, tgt), (i, j))
+    check_with_ref(d, model, value, (i, j), (a, b))
+end
+function check_with_ref(d::EdgeField, against, value, (a, b)::EdgeLabel)
+    model = get_model(against)
+    network = NF.network(model)
+    src, tgt = indexes(d, network)
+    (i, j) = N.to_index.((src, tgt), (a, b))
+    check_with_ref(d, model, value, (i, j), (a, b))
+end
+
+#-------------------------------------------------------------------------------------------
+# Construct.
+
+function construct(d::EdgeField, ::Type{<:EdgeFieldRawBlueprint}, raw)
+    T = D.type(d)
+    try
+        v = inputconvert(Vector{T}, raw)
+        for (i, value) in enumerate(v)
+            check_with_raw_ref(d, value, i)
+        end
+        v
+    catch e
+        e isa F.InputError || rethrow(e)
+        with_context!(e, "When constructing $d from raw values")
+    end
+end
+
+function construct(d::EdgeField, Field::Component, input; _...)
+    parsed = parse(d, input)
+    construct_from_parsed(d, Field, parsed)
+end
+
+construct_from_parsed(::EdgeField, Field::Component, raw::Vector) = Field.Raw(raw)
+
+function parse(d::EdgeField, input)
+    T = D.type(d)
+    tries = []
+    if may_flat(d)
+        push!(tries, T)
+    end
+    push!(tries, Vector{T})
+    M = D.is_sparse(d) ? SparseMatrix : Matrix
+    push!(tries, M{T})
+    push!(tries, Adjacency{T})
+    try_convert(input, tries...)
 end